用户授权数据同步任务修改

+++ rosettaflow-common/src/main/java/com/moirae/rosettaflow/common/constants/SysConfig.java
@@ -16,6 +16,6 @@ public class SysConfig {
     private boolean kickMode = true;
     private long nonceTimeOut = 5 * 60 * 1000;
     private String algorithmFilepath = "classpath:script/algorithm.csv";
-    private long batchSize = 500;
+    private int batchSize = 500;
     private long redisTimeOut = 24 * 60 * 60 * 1000;
 }
